% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credInt.R
\name{credInt}
\alias{credInt}
\title{Credible Interval for intended coverage}
\usage{
credInt(X, y, alpha, num_mcmc = 2000)
}
\arguments{
\item{X}{Design matrix}

\item{y}{Response variable}

\item{alpha}{Intended coverage}

\item{num_mcmc}{Number of posterior samples (default = 2000)}
}
\value{
Credible interval for all regression covariates and posterior samples
}
\description{
Credible Interval for intended coverage
}
\examples{
credInt(X,y,alpha = 0.05,num_mcmc = 5000)

}
\author{
Samhita Pal
}
