% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solveGamma.R
\name{solve_gamma}
\alias{solve_gamma}
\title{Title}
\usage{
solve_gamma(lambda, alpha)
}
\arguments{
\item{lambda}{The cross-validated penalty parameter}

\item{alpha}{The desired level of confidence}
}
\value{
The choice of gamma for credible intervals
}
\description{
Title
}
\examples{
solve_gamma(lambda = 0.02, alpha = 0.05)

}
\author{
Samhita Pal
}
